+++ src/pages/Home.jsx
@@ -5,29 +5,14 @@ import Landing from "../components/home/Landing";
 import LandingIntro from "../components/home/LandingIntro";
 import NewItems from "../components/home/NewItems";
 import TopSellers from "../components/home/TopSellers";
-import axios from "axios";
 
-const Home = () => {
-  const [collectionsData, setCollectionsData] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
+const Home = ({collectionsData, loadingState}) => {
+  console.log("home collections Data:");
+  console.log(collectionsData);
 
-  async function fetchCollections() {
-    setIsLoading(true);
-    axios.get("https://us-central1-nft-cloud-functions.cloudfunctions.net/hotCollections")
-      .then((response) => {
-         setCollectionsData(response.data)
-      })
-      .catch((error) => {
-        console.log(error);
-      })
-      .finally(() =>{
-        setIsLoading(false);
-      }
-      )
-  }
+ 
   useEffect(() => {
     window.scrollTo(0, 0);
-    fetchCollections()
   }, []);
 
   return (
@@ -36,7 +21,7 @@ const Home = () => {
         <div id="top"></div>
         <Landing />
         <LandingIntro />
-        <HotCollections collections={collectionsData}/>
+        <HotCollections collections={collectionsData} loadingState={loadingState}/>
         <NewItems />
         <TopSellers />
         <BrowseByCategory />
